+++ src/OpenTelemetry/Logs/ILogger/OpenTelemetryLogger.cs
@@ -16,7 +16,6 @@
 
 #nullable enable
 
-using System.ComponentModel;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
@@ -207,6 +206,7 @@ internal sealed class OpenTelemetryLogger : ILogger
             // Note: This is to preserve legacy behavior where State is
             // exposed if we didn't parse state.
             iLoggerData.State = state;
+
             return null;
         }
         else
@@ -215,36 +215,9 @@ internal sealed class OpenTelemetryLogger : ILogger
             // have come from the pool and may have State from a prior log.
             iLoggerData.State = null;
 
-            try
-            {
-                PropertyDescriptorCollection itemProperties = TypeDescriptor.GetProperties(state!);
+            OpenTelemetrySdkEventSource.Log.LoggerProcessStateSkipped<TState>();
 
-                var attributeStorage = logRecord.AttributeStorage ??= new List<KeyValuePair<string, object?>>(itemProperties.Count);
-
-                foreach (PropertyDescriptor? itemProperty in itemProperties)
-                {
-                    if (itemProperty == null)
-                    {
-                        continue;
-                    }
-
-                    object? value = itemProperty.GetValue(state);
-                    if (value == null)
-                    {
-                        continue;
-                    }
-
-                    attributeStorage.Add(new KeyValuePair<string, object?>(itemProperty.Name, value));
-                }
-
-                return attributeStorage;
-            }
-            catch (Exception parseException)
-            {
-                OpenTelemetrySdkEventSource.Log.LoggerParseStateException<TState>(parseException);
-
-                return Array.Empty<KeyValuePair<string, object?>>();
-            }
+            return Array.Empty<KeyValuePair<string, object?>>();
         }
     }