+++ src/OpenTelemetry/CHANGELOG.md
@@ -8,6 +8,16 @@
   `IReadOnlyList` or `IEnumerable` of `KeyValuePair<string, object>`s.
   ([#4609](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4609))
 
+* **Breaking Change** Removed the support for parsing `TState` types passed to
+  the `ILogger.Log<TState>` API when `ParseStateValues` is true and `TState`
+  does not implement either `IReadOnlyList<KeyValuePair<string, object>>` or
+  `IEnumerable<KeyValuePair<string, object>>`. This feature was first introduced
+  in the `1.5.0` stable release with
+  [#4334](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4334) and
+  has been removed because it makes the OpenTelemetry .NET SDK incompatible with
+  native AOT.
+  ([#4614](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4614))
+
 ## 1.5.0
 
 Released 2023-Jun-05